Refactor to use linkField
