+++ src/payload-types.ts
@@ -1835,6 +1835,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -1850,12 +1851,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -1871,7 +1872,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -1892,6 +1892,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -1907,12 +1908,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -1928,7 +1929,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -1955,6 +1955,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -1970,12 +1971,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -1991,7 +1992,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -2018,6 +2018,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -2033,12 +2034,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -2054,7 +2055,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -2081,6 +2081,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -2096,12 +2097,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -2117,7 +2118,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -2144,6 +2144,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -2159,12 +2160,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -2180,7 +2181,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -2207,6 +2207,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -2222,12 +2223,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -2243,7 +2244,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -2270,6 +2270,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -2285,12 +2286,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -2306,7 +2307,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -2333,6 +2333,7 @@ export interface Navigation {
           label?: string | null;
           link?: {
             type?: ('internal' | 'external') | null;
+            newTab?: boolean | null;
             reference?:
               | ({
                   relationTo: 'pages';
@@ -2348,12 +2349,12 @@ export interface Navigation {
                 } | null);
             url?: string | null;
             label?: string | null;
-            newTab?: boolean | null;
           };
           items?:
             | {
                 link?: {
                   type?: ('internal' | 'external') | null;
+                  newTab?: boolean | null;
                   reference?:
                     | ({
                         relationTo: 'pages';
@@ -2369,7 +2370,6 @@ export interface Navigation {
                       } | null);
                   url?: string | null;
                   label?: string | null;
-                  newTab?: boolean | null;
                 };
                 id?: string | null;
               }[]
@@ -2384,6 +2384,7 @@ export interface Navigation {
     };
     link?: {
       type?: ('internal' | 'external') | null;
+      newTab?: boolean | null;
       reference?:
         | ({
             relationTo: 'pages';
@@ -2399,7 +2400,6 @@ export interface Navigation {
           } | null);
       url?: string | null;
       label?: string | null;
-      newTab?: boolean | null;
     };
   };
   contentHash?: string | null;
@@ -3492,10 +3492,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3504,10 +3504,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3525,10 +3525,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3537,10 +3537,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3563,10 +3563,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3575,10 +3575,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3601,10 +3601,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3613,10 +3613,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3639,10 +3639,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3651,10 +3651,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3677,10 +3677,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3689,10 +3689,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3715,10 +3715,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3727,10 +3727,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3753,10 +3753,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3765,10 +3765,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3791,10 +3791,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                 | T
                 | {
                     type?: T;
+                    newTab?: T;
                     reference?: T;
                     url?: T;
                     label?: T;
-                    newTab?: T;
                   };
               items?:
                 | T
@@ -3803,10 +3803,10 @@ export interface NavigationsSelect<T extends boolean = true> {
                       | T
                       | {
                           type?: T;
+                          newTab?: T;
                           reference?: T;
                           url?: T;
                           label?: T;
-                          newTab?: T;
                         };
                     id?: T;
                   };
@@ -3825,10 +3825,10 @@ export interface NavigationsSelect<T extends boolean = true> {
           | T
           | {
               type?: T;
+              newTab?: T;
               reference?: T;
               url?: T;
               label?: T;
-              newTab?: T;
             };
       };
   contentHash?: T;
